[build] do not depends on autoconf-archive to reconfigure libffi
libffi uses the AX_CHECK_COMPILE_FLAG that is defined into the autoconf-archive package.
We put the macro into local macros dir so that we do not need autoconf-archive installed
when we run bootstrap.

Note that manually running autoreconf (instead of running bootstrap) needs either:
- [A] autoconf-archive to be installed
- [B] ACLOCAL_AMFLAGS including ${top_srcdir}/acinclude
- [C] ACLOCAL including ${top_srcdir}/acinclude
B is what is done in bootstrap.