[tests] fix an error in the atlocal.in file
We need the docdir value to be able to run installed tests (installed examples)
So, we substitute the value with autoconf.
However, we also need to compute shell variables (recursively) included into the docdir value.

datarootdir was not computed but was fixed by a autoconf workaround for this variable.
Better correctly fix this locally.