[kaapic/kaapif] fix fortran support for spawning tasks
Note: as body and nargs parameters switch from the previous implementation,
this allows to write a error message and to abord if nargs>100.
This is probably due to the fact that body and nargs was not switched
and also mode, value, count, type permuted to mode, type, count, value
(there is no way to automaticcaly detect this last part).