[linbgomp] libkomp-run: remove --gdb and --time support
Let libkomp-run do its minimum work.
It is as easy to call "libkomp-run --time ..." than "time libkomp-run ..."
gdb support was broken:
- write in pkglibdir at runtime (usually a read-only location)
- do not handle correctly program arguments

gdb use the environment variable from the current environment.
However, as stated in its doc:
   _Warning:_ On Unix systems, GDB runs your program using the shell
indicated by your `SHELL' environment variable if it exists (or
`/bin/sh' if not).  If your `SHELL' variable names a shell that runs an
initialization file--such as `.cshrc' for C-shell, or `.bashrc' for
BASH--any variables you set in that file affect your program.  You may
wish to move setting of environment variables to files that are only
run when you sign on, such as `.login' or `.profile'.

  So, in case of problem with gdb, check your startup files